pyproject.toml
[project]
name = "case-mle-picpay"
version = "0.1.0"
description = "The repository stores the code related to the challenge case for the full-time MLE position at PicPay."
authors = [
    {name = "jordaoa",email = "andre.nascimento@ccc.ufcg.edu.br"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pyspark (>=4.1.1,<5.0.0)",
    "setuptools (>=65.0.0,<70.0.0)",
    "numpy (<2.0.0)",
    "mlflow (>=2.13.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
